Move __mocks__ package out of __tests__
